Move locktime and sequence code to units
Move the `sequence` module and the code from both locktime modules
from `primitives` to `units`.
