+++ bitcoin/src/blockdata/mod.rs
@@ -39,9 +39,9 @@ pub mod locktime {
 
         pub use crate::consensus::encode::{self, Decodable, Encodable};
 
-        /// Re-export everything from the `primitives::locktime::absolute` module.
+        /// Re-export everything from the `units::locktime::absolute` module.
         #[rustfmt::skip]        // Keep public re-exports separate.
-        pub use primitives::locktime::absolute::{ConversionError, Height, LockTime, ParseHeightError, ParseTimeError, MedianTimePast};
+        pub use units::locktime::absolute::{ConversionError, Height, LockTime, ParseHeightError, ParseTimeError, MedianTimePast};
 
         #[deprecated(since = "TBD", note = "use `MedianTimePast` instead")]
         #[doc(hidden)]
@@ -69,8 +69,8 @@ pub mod locktime {
         //! There are two types of lock time: lock-by-height and lock-by-time, distinguished by
         //! whether bit 22 of the `u32` consensus value is set.
 
-        /// Re-export everything from the `primitives::locktime::relative` module.
-        pub use primitives::locktime::relative::{
+        /// Re-export everything from the `units::locktime::relative` module.
+        pub use units::locktime::relative::{
             DisabledLockTimeError, InvalidHeightError, InvalidTimeError, LockTime,
             NumberOf512Seconds, NumberOfBlocks, TimeOverflowError,
         };
