+++ primitives/src/lib.rs
@@ -38,13 +38,11 @@ pub mod _export {
 }
 
 pub mod block;
-pub mod locktime;
 pub mod merkle_tree;
 mod opcodes;
 pub mod pow;
 #[cfg(feature = "alloc")]
 pub mod script;
-pub mod sequence;
 pub mod transaction;
 #[cfg(feature = "alloc")]
 pub mod witness;
@@ -54,6 +52,8 @@ pub use units::{
     amount::{self, Amount, SignedAmount},
     block::{BlockHeight, BlockHeightInterval, BlockMtp, BlockMtpInterval},
     fee_rate::{self, FeeRate},
+    locktime::{self, absolute, relative},
+    sequence::{self, Sequence},
     time::{self, BlockTime},
     weight::{self, Weight},
 };
@@ -75,10 +75,8 @@ pub use self::{
 #[doc(inline)]
 pub use self::{
     block::{BlockHash, Header as BlockHeader, Version as BlockVersion, WitnessCommitment},
-    locktime::{absolute, relative},
     merkle_tree::{TxMerkleNode, WitnessMerkleNode},
     pow::CompactTarget,
-    sequence::Sequence,
     transaction::{OutPoint, Txid, Version as TransactionVersion, Wtxid},
 };
