--- primitives/src/locktime/mod.rs
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: CC0-1.0
-
-//! Provides absolute and relative locktimes.
-
-pub mod absolute;
-pub mod relative;
