+++ primitives/src/transaction.rs
@@ -24,19 +24,19 @@ use hashes::sha256d;
 use internals::compact_size;
 #[cfg(feature = "hex")]
 use internals::write_err;
+#[cfg(feature = "alloc")]
+use units::locktime::absolute;
 #[cfg(feature = "hex")]
 use units::parse;
+#[cfg(feature = "alloc")]
+use units::sequence::Sequence;
 #[cfg(feature = "alloc")]
 use units::{Amount, Weight};
 
 #[cfg(feature = "alloc")]
-use crate::locktime::absolute;
-#[cfg(feature = "alloc")]
 use crate::prelude::Vec;
 #[cfg(feature = "alloc")]
 use crate::script::ScriptBuf;
-#[cfg(feature = "alloc")]
-use crate::sequence::Sequence;
 #[cfg(feature = "alloc")]
 use crate::witness::Witness;
