+++ units/Cargo.toml
@@ -20,7 +20,7 @@ alloc = ["internals/alloc","serde?/alloc"]
 [dependencies]
 internals = { package = "bitcoin-internals", path = "../internals", version = "0.4.0" }
 
-serde = { version = "1.0.103", default-features = false, optional = true }
+serde = { version = "1.0.103", default-features = false, features = ["derive"], optional = true }
 arbitrary = { version = "1.4", optional = true }
 
 [dev-dependencies]
